Capitalise 'Perl' when meaning the language
The language 'Perl' us supposed to be capitalised.  When referring to
the interpreter, then one should use the lower case version 'perl'.
I've tried to update the docs here consistently so that when talking
about Perl in the more general, language-related sense, the word has
been capitalised.  However, when talking about the "perl process" I've
left the word in lower case since this really relates to the
interpreter.